Feature:add edit subProject fix: Zero Qodana problems Fix: test fix fix: Front page picture add fix:Scroll function on my-projects.html on smaller screens
