+++ src/main/java/com/example/alphasolutionsv2/service/RoleService.java
@@ -5,6 +5,7 @@ import com.example.alphasolutionsv2.repository.RoleRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class RoleService {
