+++ src/main/resources/templates/my-projects.html
@@ -4,6 +4,12 @@
   <title>Mine Projekter</title>
   <!-- Først Bootstrap -->
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+  <style>
+    .table-responsive{
+      overflow-x: auto;
+      -webkit-overflow-scrolling: touch;
+    }
+  </style>
 </head>
 <body>
 
@@ -33,7 +39,7 @@
   <div th:if="${projects.isEmpty()}">
     <div class="alert alert-info">Du har endnu ikke oprettet nogen projekter.</div>
   </div>
-
+<div class="table-responsive">
   <table class="table table-striped" th:unless="${projects.isEmpty()}">
     <thead>
     <tr>
@@ -84,6 +90,7 @@
   </div>
   <!-- SLUT ADMIN SEKTION -->
 
+</div>
 </div>
 </body>
 </html>
